microservices/dashboard/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

psycopg2-binary==2.9.9

pydantic==2.6.1
python-dotenv==1.0.1
